+++ app/Http/Controllers/API/AuthController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
+use Laravel\Passport\Passport;
 use App\Models\User;
 
-class RegisterController extends Controller
+class AuthController extends Controller
 {
     public function register(Request $request){
 
@@ -29,4 +31,25 @@ class RegisterController extends Controller
         ]);
         
     }
+
+    public function login(Request $request){
+        $loginData =$request->validate([
+            'email'=> 'email|required',
+            'password' => 'required'
+        ]);
+        if(!Auth()->attempt($loginData)){
+            return response([
+                'message' => 'Invalid Credentials',
+            ]);
+        }
+
+            $user = $request->user();
+            $accessToken = $user->createToken('authToken')->accessToken;
+
+            return response(['user'=> Auth::user(), 'access_token'=> $accessToken, 'status' => 200]
+        );
+          
+    }
+
 }
+
